requirements.txt
numpy
pillow
tensorflow
streamlit
streamlit_extras
